feature/INTAP-1722 add api
