+++ src/Training/Bundle/UserNamingBundle/Resources/config/oro/api.yml
@@ -1,0 +1,15 @@
+api:
+  entities:
+    Training\Bundle\UserNamingBundle\Entity\UserNamingType:
+      sorters:
+        fields:
+          title:
+            exclude: false
+      filters:
+        fields:
+          title:
+            exclude: false
+      fields:
+        nameExample:
+          data_type: string
+          meta_property: true
