+++ src/Training/Bundle/UserNamingBundle/Twig/UserNamingExtension.php
@@ -10,33 +10,13 @@ use Twig\TwigFilter;
 
 class UserNamingExtension extends AbstractExtension
 {
-    public function __construct(private UserFullNameProvider $fullNameProvider)
-    {
-    }
-
     /**
      * {@inheritdoc}
      */
     public function getFilters(): array
     {
         return [
-            new TwigFilter('full_name_filter', [$this, 'getFullNameForFilter'])
+            new TwigFilter('full_name_example', [UserFullNameProvider::class, 'getFullNameExample'])
         ];
-    }
-
-    /**
-     * @param string $format
-     * @return string
-     */
-    public function getFullNameForFilter(string $format): string
-    {
-        $user = new User();
-        $user->setNamePrefix('User')
-            ->setFirstName('John')
-            ->setMiddleName('Jon')
-            ->setLastName('Doe')
-            ->setNameSuffix('Sr');
-
-        return $this->fullNameProvider->getFullName($user, $format);
     }
 }
